model_training/whisper/pyproject.toml
[tool.poetry]
name = "whisper"
version = "0.1.0"
description = ""
authors = ["Kenny Li <karlili.li@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wandb = "^0.16.6"
dataset = "^1.6.2"
transformers = "^4.39.3"
evaluate = "^0.4.1"
librosa = "^0.10.1"
soundfile = "^0.12.1"
torch = "^2.2.2"
ffmpeg-python = "^0.2.0"
jiwer = "^3.0.3"
accelerate = "^0.29.1"
pandas = "^2.2.2"
tensorboardx = "^2.6.2.2"
pycantonese = "^3.4.0"
jupyterlab = "^4.2.0"
pyannote-audio = "^3.2.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
